+++ java/core/clustercontroller/src/test/java/io/joynr/messaging/LibJoynrMessageRouterTest.java
@@ -87,11 +87,11 @@ public class LibJoynrMessageRouterTest {
         message.setExpirationDate(ExpiryDate.fromRelativeTtl(10000));
         message.setTo(unknownParticipantId);
         message.setLocalMessage(false);
-        message.setType(message.MESSAGE_TYPE_REQUEST);
+        message.setType(JoynrMessage.MESSAGE_TYPE_REQUEST);
 
         when(routingTable.containsKey(unknownParticipantId)).thenReturn(false);
         when(messageRouterParent.resolveNextHop(unknownParticipantId)).thenReturn(true);
-        when(messageRouterParent.getGlobalAddress()).thenReturn(globalAddress);
+        when(messageRouterParent.getReplyToAddress()).thenReturn(globalAddress);
         when(messagingStubFactory.create(Mockito.any(Address.class))).thenReturn(messagingStub);
         when(parentAddress.getChannelId()).thenReturn("LibJoynrMessageRouterTestChannel");
 
@@ -131,7 +131,6 @@ public class LibJoynrMessageRouterTest {
     }
 
     @Test
-    @SuppressWarnings("unchecked")
     public void passesNextHopToParent() {
         messageRouter.addNextHop(unknownParticipantId, nextHopAddress);
         Mockito.verify(messageRouterParent).addNextHop(Mockito.eq(unknownParticipantId), Mockito.eq(incomingAddress));