+++ java/core/libjoynr/src/main/java/io/joynr/messaging/routing/LibJoynrMessageRouter.java
@@ -196,8 +196,7 @@ public class LibJoynrMessageRouter extends AbstractMessageRouter {
         }
         deferredParentHopsParticipantIds.clear();
 
-        String globalAddress = parentRouter.getGlobalAddress();
-        replyToAddress = globalAddress;
+        replyToAddress = parentRouter.getReplyToAddress();
     }
 
     /**